Change: move create_note to dedicated files
