+++ src/manage_sql_notes.h
@@ -1,0 +1,12 @@
+/* Copyright (C) 2026 Greenbone AG
+ *
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+#ifndef _GVMD_MANAGE_SQL_NOTES_H
+#define _GVMD_MANAGE_SQL_NOTES_H
+
+#include "manage_notes.h"
+
+
+#endif // not _GVMD_MANAGE_SQL_NOTES_H
